src/main/java/it/unipd/tos/business/TakeAwayBillClass.java
////////////////////////////////////////////////////////////////////
// [MASSIMO] [TOFFOLETTO] [1161727]
////////////////////////////////////////////////////////////////////
package it.unipd.tos.business;

import java.util.List;

import it.unipd.tos.business.exception.TakeAwayBillException;
import it.unipd.tos.model.MenuItem;

public class TakeAwayBillClass implements TakeAwayBill {

    public double getOrderPrice(List<MenuItem> itemsOrdered) 
            throws TakeAwayBillException {
        double totPrice=0.0d;
        for(MenuItem item:itemsOrdered){
            totPrice+=item.getPrice();
        }
        return totPrice;
    }

}
